refactor(assistant): move delivery/media/event stores into persistence/ with shims
